Moved map control files.
