+++ src/app/layout/datasets/openlayermap/controls/olmap.zoom.component.ts
